+++ src/main/java/com/feed_the_beast/ftbquests/quest/QuestChapter.java
@@ -128,9 +128,9 @@ public final class QuestChapter extends ProgressingQuestObject
 	}
 
 	@Override
-	public int getProgress(IProgressData data)
+	public long getProgress(IProgressData data)
 	{
-		int progress = 0;
+		long progress = 0L;
 
 		for (Quest quest : quests)
 		{
@@ -141,9 +141,9 @@ public final class QuestChapter extends ProgressingQuestObject
 	}
 
 	@Override
-	public int getMaxProgress()
+	public long getMaxProgress()
 	{
-		int maxProgress = 0;
+		long maxProgress = 0L;
 
 		for (Quest quest : quests)
 		{
